+++ frontend/src/pages/AppPage/Tasks/TaskForm.scss
@@ -79,10 +79,6 @@
     border: none;
     cursor: pointer;
     transition: 0.5s;
-    background-color: var(--dark-gray-700);
-}
-
-.input_btn:nth-child(2) {
     background-color: var(--primary-400);
-    margin: 0 0 0 10px;
+
 }
