+++ frontend/src/pages/AppPage/Tasks/Tasks.tsx
@@ -1,53 +1,15 @@
 import React, { useEffect } from 'react';
 import './tasks.scss';
-import { FormikHelpers, useFormik } from 'formik';
-import getTaskSchema from '../../../schemas/taskSchema';
 import useTodos from '../../../hooks/useTodos';
-import useProjects from '../../../hooks/useProjects';
 import TaskElem from './TaskElem';
-import TaskForm, { TaskFormValues } from '../TaskForm/TaskForm';
 import useAuth from '../../../hooks/useAuth';
-import useContexts from '../../../hooks/useContexts';
+import AddTask from './AddTask';
 
 const Tasks = () => {
-  const { currentProjectId } = useProjects();
-  const { currentContextId } = useContexts();
-  const { currentListViewId } = useTodos();
   const {
-    fetchTodos, addNewTodo, todos, currentTitle,
+    fetchTodos, todos, currentTitle,
   } = useTodos();
   const { getHeaders } = useAuth();
-
-  const onSubmit = async (body: TaskFormValues, formikHelpers: FormikHelpers<TaskFormValues>) => {
-    const { resetForm } = formikHelpers;
-    const { title, projectId, contextId } = body;
-    addNewTodo({
-      createTodoDto: {
-        title,
-        completed: false,
-        projectId: projectId === undefined ? null : projectId,
-        contextId: contextId === undefined ? null : contextId,
-      },
-      headers: getHeaders(),
-    });
-    resetForm();
-  };
-
-  const initialValues: TaskFormValues = {
-    title: '',
-    description: '',
-    projectId: currentProjectId !== null
-      && currentProjectId === currentListViewId ? currentProjectId : undefined,
-    contextId: currentContextId !== null
-      && currentContextId === currentListViewId ? currentContextId : undefined,
-  };
-
-  const form = useFormik({
-    initialValues,
-    onSubmit,
-    validationSchema: getTaskSchema(),
-    enableReinitialize: true,
-  });
 
   useEffect(() => {
     fetchTodos({ headers: getHeaders() });
@@ -61,7 +23,7 @@ const Tasks = () => {
           <TaskElem task={taskProps} />
         </React.Fragment>
       ))}
-      <TaskForm form={form} />
+      <AddTask />
     </div>
   );
 };
