+++ frontend/src/schemas/taskSchema.ts
@@ -2,9 +2,8 @@ import * as yup from 'yup';
 
 const getTaskSchema = () => yup.object().shape({
   title: yup.string().required().max(144),
-  description: yup.string().max(255),
-  projectId: yup.number().nullable(),
-  contextId: yup.number().nullable(),
+  projectId: yup.string(),
+  contextId: yup.string(),
 });
 
 export default getTaskSchema;
