+++ frontend/src/types/taskFormType.ts
@@ -1,0 +1,5 @@
+export type CreateTodoForm = {
+  title: string,
+  projectId: string | 'default';
+  contextId: string | 'default';
+};
